backend/requirements.txt
djangorestframework
Pillow
reportlab
PyMuPDF
django-cors-headers
